Exp: Explore Args implementation using clap's derive features.
I thought this was going to be a simple commit, but changing how args work
pushed me into making more changes than I anticipated,

  - Unifying `url_download` and `list_download` functions
    - Extending save-as flag to work with multiple downloads
  - Ultimately removing `src/parser.rs` as I needed to dismantle,
    - Manual argument definition
    - Non-parsing work, like the implementation for all sub-commands
  - Gathering more packages,
    - `clap`'s derive argument
    - `carapace_spec_clap` to generate autocompletion
      - There's `clap_complete` too, I'm not sure which one to use.
    - `blake3` to avoid indexing with integers and making urls order dependent
  - This also bumped the lockfile to v4 too. It's a minor update, but does not
    belong on this commit

I'll leave this commit in a experimental branch and work on a series of bite
sized commits to get here. I don't want commits to be hard to review as it leads
to sloppy code.