Misc: Switch from directories to dirs
[directories](https://docs.rs/directories) tries to be friendly by collecting all the
information on one go, but fails to be friendly when defaults are needed.

On another machine rawst crashes on this unwrap, which is quite unexpected as
most tools assume ~/Downloads/ as the download directory in case not all the XDG
environment variables are set up.

```rust
    download_dir: user_dirs.download_dir().unwrap().to_path_buf(),
```

With this in mind, I looked for another library, and
[dirs](https://docs.rs/dirs) seems to try harder on a fallback while also being
more minimalistic.